backend/requirements.txt
Django==4.2.10
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-allauth==0.57.0
django-cors-headers==4.3.0
psycopg2-binary==2.9.9
qrcode==7.4.2
django-storages==1.14.2
boto3==1.28.64
Pillow==10.0.1
pytest==7.4.2
pytest-django==4.5.2
factory-boy==3.3.0
gunicorn==21.2.0
python-dotenv==1.0.0 